Added support for negative ints
