+++ src/lexer/mod.rs
@@ -1,11 +1,13 @@
-pub mod common;
 pub mod error;
+pub mod identifier;
 pub mod keyword;
+pub mod literal;
 pub mod operator;
 pub mod token;
 
 pub use error::LexerError;
 
-pub use common::{lex_identifier, lex_literal, Identifier, Literal};
+pub use identifier::{lex_identifier, Identifier};
 pub use keyword::{lex_keyword, Keyword};
+pub use literal::{lex_literal, Literal};
 pub use operator::{lex_operator, Operator};
